# Conflicts:
#	EmploymentWorks.vcxproj
#	EmploymentWorks.vcxproj.filters
#	KazLibrary/Buffer/DescriptorHeapMgr.cpp
#	KazLibrary/Buffer/DescriptorHeapMgr.h
#	KazLibrary/DirectXCommon/Base.h
#	KazLibrary/Helper/KazBufferHelper.cpp
#	KazLibrary/Helper/KazBufferHelper.h
#	KazLibrary/Math/KazMath.cpp
#	KazLibrary/Pipeline/GraphicsPipeLineMgr.cpp
#	KazLibrary/Pipeline/GraphicsPipeLineMgr.h
#	KazLibrary/Pipeline/GraphicsRootSignature.cpp
#	KazLibrary/Pipeline/GraphicsRootSignature.h
#	KazLibrary/Pipeline/PreCreateBasePipeLine.cpp
#	KazLibrary/Pipeline/Shader.h
#	KazLibrary/Render/BoxPolygonRender.cpp
#	KazLibrary/Render/LineRender.cpp
#	KazLibrary/RenderTarget/RenderTargetStatus.h
#	KazLibrary/Scene/ClassScene.cpp
#	KazLibrary/Scene/ClassScene.h
#	KazLibrary/Scene/DebugScene.cpp
#	KazLibrary/Scene/DebugScene.h
#	KazLibrary/Scene/SceneManager.cpp
#	KazLibrary/Scene/main.cpp
#	Resource/ShaderFiles/ComputeShader/TestComputeShader.hlsl

+++ KazLibrary/Helper/KazHelper.cpp
@@ -144,7 +144,7 @@ std::vector<char> KazHelper::CovertStringToChar(std::string STRING)
 }
 
 
-void KazHelper::ConvertStringToWchar_t(string STRING, wchar_t *WCHAR_STRING, int ARRAY_SIZE)
+void KazHelper::ConvertStringToWchar_t(string STRING, wchar_t *WCHAR_STRING, size_t ARRAY_SIZE)
 {
-	MultiByteToWideChar(CP_ACP, 0, STRING.c_str(), -1, WCHAR_STRING, ARRAY_SIZE);
+	MultiByteToWideChar(CP_ACP, 0, STRING.c_str(), -1, WCHAR_STRING, static_cast<INT>(ARRAY_SIZE));
 }
